requirements.txt
langchain
langchain-community
sentence-transformers
faiss-cpu
google-generativeai
transformers
numpy
torch
python-dotenv
python-dotenv